backend/requirements.txt
Django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.6.*
psycopg2-binary==2.9.*
gunicorn==23.*
celery==5.4.*
redis==5.2.*
django-celery-beat==2.7.*
google-auth==2.36.*
google-auth-oauthlib==1.2.*
google-api-python-client==2.155.*
dj-database-url==2.3.*
python-decouple==3.8
